trans scene

+++ game/MMO/NFServerLogic/NFGameServer/NFBattlePlugin/Creature/NFCreatureVisionData.h
@@ -64,7 +64,7 @@ public:
     int ResumeInit();
 
 public:
-    int creatureCid;
+    uint64_t creatureCid;
     int nMeInHisVisionPos;
 };
